+++ src/api/healthCheck.ts
@@ -22,11 +22,11 @@ const healthStatusRoute = (server: FastifySeverInstance) => {
       },
     },
     async (request, reply) => {
-      const result = await (
-        await (await getHealthCheck()).getHealthStatus()
-      ).map((status) => ({
+      const health = await getHealthCheck();
+      const result = (await health.getHealthStatus()).map((status) => ({
         ...status,
         lastCheck: status.lastCheck?.toISOString(),
+        lastTrialErrorTime: status.lastTrialErrorTime?.toISOString(),
       }));
       reply.status(200).send(result);
     }
@@ -51,14 +51,18 @@ const healthStatusForParameterRoute = (server: FastifySeverInstance) => {
     },
     async (request, reply) => {
       const health = await getHealthCheck();
-      const status = await health.getHealthStatusFor(request.params.paramId);
+      const status = await health.getHealthStatusWithParamId(
+        request.params.paramId
+      );
       if (!status)
         throw new Error(
           `Health parameter with id '${request.params.paramId}' is not registered.`
         );
-      reply
-        .status(200)
-        .send({ ...status, lastCheck: status.lastCheck!.toISOString() });
+      reply.status(200).send({
+        ...status,
+        lastCheck: status.lastCheck?.toISOString(),
+        lastTrialErrorTime: status.lastTrialErrorTime?.toISOString(),
+      });
     }
   );
 };
@@ -83,14 +87,18 @@ const updateHealthStatusForParameterRoute = (server: FastifySeverInstance) => {
     async (request, reply) => {
       const health = await getHealthCheck();
       await health.updateParam(request.params.paramId);
-      const status = await health.getHealthStatusFor(request.params.paramId);
+      const status = await health.getHealthStatusWithParamId(
+        request.params.paramId
+      );
       if (!status)
         throw new Error(
           `Health parameter with id '${request.params.paramId}' is not registered.`
         );
-      reply
-        .status(200)
-        .send({ ...status, lastCheck: status.lastCheck!.toISOString() });
+      reply.status(200).send({
+        ...status,
+        lastCheck: status.lastCheck?.toISOString(),
+        lastTrialErrorTime: status.lastTrialErrorTime?.toISOString(),
+      });
     }
   );
 };