+++ src/guard/HealthCheck.ts
@@ -4,6 +4,7 @@ import {
   CardanoKoiosAssetHealthCheckParam,
   ErgoExplorerAssetHealthCheckParam,
   ErgoNodeAssetHealthCheckParam,
+  EthereumRpcAssetHealthCheckParam,
 } from '@rosen-bridge/asset-check';
 import { HealthCheck, HealthStatusLevel } from '@rosen-bridge/health-check';
 import { LogLevelHealthCheck } from '@rosen-bridge/log-level-check';
@@ -12,6 +13,7 @@ import { P2PNetworkHealthCheck } from '@rosen-bridge/p2p-network-check';
 import {
   ErgoExplorerScannerHealthCheck,
   ErgoNodeScannerHealthCheck,
+  EthereumRPCScannerHealthCheck,
 } from '@rosen-bridge/scanner-sync-check';
 
 import WinstonLogger from '@rosen-bridge/winston-logger';
@@ -26,11 +28,19 @@ import Configs from '../configs/Configs';
 import GuardsCardanoConfigs from '../configs/GuardsCardanoConfigs';
 import GuardsErgoConfigs from '../configs/GuardsErgoConfigs';
 import { rosenConfig } from '../configs/RosenConfig';
-import { dataSource } from '../db/dataSource';
 import GuardPkHandler from '../handlers/GuardPkHandler';
 import { ADA_DECIMALS, ERG_DECIMALS } from '../utils/constants';
 import GuardsBitcoinConfigs from '../configs/GuardsBitcoinConfigs';
-import { BITCOIN_CHAIN } from '@rosen-chains/bitcoin';
+import { BITCOIN_CHAIN, BTC } from '@rosen-chains/bitcoin';
+import { DatabaseAction } from '../db/DatabaseAction';
+import { NotFoundError } from '@rosen-chains/abstract-chain';
+import { NotificationHandler } from '../handlers/NotificationHandler';
+import {
+  TxInfo,
+  TxProgressHealthCheckParam,
+} from '@rosen-bridge/tx-progress-check';
+import GuardsEthereumConfigs from '../configs/GuardsEthereumConfigs';
+import { ETH, ETHEREUM_CHAIN } from '@rosen-chains/ethereum';
 
 const logger = WinstonLogger.getInstance().getLogger(import.meta.url);
 let healthCheck: HealthCheck | undefined;
@@ -41,18 +51,49 @@ let healthCheck: HealthCheck | undefined;
  */
 const getHealthCheck = async () => {
   if (!healthCheck) {
-    healthCheck = new HealthCheck();
+    // initialize HealthCheck
+    const notificationHandler = NotificationHandler.getInstance();
+    const notificationConfig = {
+      historyConfig: {
+        cleanupThreshold: Configs.historyCleanupThreshold,
+      },
+      notificationCheckConfig: {
+        hasBeenUnstableForAWhile: {
+          windowDuration: Configs.hasBeenUnstableForAWhileWindowDuration,
+        },
+        hasBeenUnknownForAWhile: {
+          windowDuration: Configs.hasBeenUnknownForAWhileWindowDuration,
+        },
+        isStillUnhealthy: {
+          windowDuration: Configs.isStillUnhealthyWindowDuration,
+        },
+      },
+    };
+    healthCheck = new HealthCheck(
+      notificationHandler.notify,
+      notificationConfig
+    );
 
+    // add LogLevel param
     const errorLogHealthCheck = new LogLevelHealthCheck(
-      logger,
+      WinstonLogger.getInstance().getDefaultLogger(),
       HealthStatusLevel.UNSTABLE,
       Configs.errorLogAllowedCount,
-      Configs.errorLogDuration,
+      Configs.logDuration,
       'error'
     );
+    const warnLogHealthCheck = new LogLevelHealthCheck(
+      WinstonLogger.getInstance().getDefaultLogger(),
+      HealthStatusLevel.UNSTABLE,
+      Configs.warnLogAllowedCount,
+      Configs.logDuration,
+      'warn'
+    );
     healthCheck.register(errorLogHealthCheck);
-    const dialer = await Dialer.getInstance();
+    healthCheck.register(warnLogHealthCheck);
 
+    // add P2PNetwork param
+    const dialer = await Dialer.getInstance();
     const p2pHealthCheck = new P2PNetworkHealthCheck({
       defectConfirmationTimeWindow: Configs.p2pDefectConfirmationTimeWindow,
       connectedGuardsHealthyThreshold:
@@ -67,9 +108,46 @@ const getHealthCheck = async () => {
     });
     healthCheck.register(p2pHealthCheck);
 
+    // add TxProgress param
+    const getActiveTransactions = async (): Promise<TxInfo[]> => {
+      return (await DatabaseAction.getInstance().getActiveTransactions()).map(
+        (txEntity) => ({
+          txId: txEntity.txId,
+          txType: txEntity.type,
+          signFailedCount: txEntity.signFailedCount,
+          chain: txEntity.chain,
+          eventId: txEntity.event.id,
+        })
+      );
+    };
+    const txProgressHealthCheck = new TxProgressHealthCheckParam(
+      getActiveTransactions,
+      Configs.txSignFailedWarnThreshold,
+      Configs.txSignFailedCriticalThreshold
+    );
+    healthCheck.register(txProgressHealthCheck);
+
     const ergoContracts = rosenConfig.contractReader(ERGO_CHAIN);
     const cardanoContracts = rosenConfig.contractReader(CARDANO_CHAIN);
     const bitcoinContracts = rosenConfig.contractReader(BITCOIN_CHAIN);
+    const ethereumContracts = rosenConfig.contractReader(ETHEREUM_CHAIN);
+
+    const generateLastBlockFetcher = (scannerName: string) => {
+      return async () => {
+        try {
+          return await DatabaseAction.getInstance().getLastSavedBlockForScanner(
+            scannerName
+          );
+        } catch (e) {
+          if (e instanceof NotFoundError) {
+            logger.info(
+              `No block found in database. Passing 0 as last height to HealthCheck`
+            );
+            return 0;
+          } else throw e;
+        }
+      };
+    };
 
     if (GuardsErgoConfigs.chainNetworkName === NODE_NETWORK) {
       const ergAssetHealthCheck = new ErgoNodeAssetHealthCheckParam(
@@ -94,9 +172,10 @@ const getHealthCheck = async () => {
       );
       healthCheck.register(emissionTokenAssetHealthCheck);
 
+      const scannerName = 'ergo-node';
       const ergoScannerSyncCheck = new ErgoNodeScannerHealthCheck(
-        dataSource,
-        'ergo-node',
+        generateLastBlockFetcher(scannerName),
+        scannerName,
         Configs.ergoScannerWarnDiff,
         Configs.ergoScannerCriticalDiff,
         GuardsErgoConfigs.node.url
@@ -135,9 +214,10 @@ const getHealthCheck = async () => {
         );
       healthCheck.register(emissionTokenAssetHealthCheck);
 
+      const scannerName = 'ergo-explorer';
       const ergoScannerSyncCheck = new ErgoExplorerScannerHealthCheck(
-        dataSource,
-        'ergo-explorer',
+        generateLastBlockFetcher(scannerName),
+        scannerName,
         Configs.ergoScannerWarnDiff,
         Configs.ergoScannerCriticalDiff,
         GuardsErgoConfigs.explorer.url
@@ -177,7 +257,7 @@ const getHealthCheck = async () => {
     }
     if (GuardsBitcoinConfigs.chainNetworkName === 'esplora') {
       const btcAssetHealthCheck = new BitcoinEsploraAssetHealthCheckParam(
-        'BTC',
+        BTC,
         bitcoinContracts.lockAddress,
         Configs.btcWarnThreshold,
         Configs.btcCriticalThreshold,
@@ -185,6 +265,30 @@ const getHealthCheck = async () => {
         8
       );
       healthCheck.register(btcAssetHealthCheck);
+    }
+    if (GuardsEthereumConfigs.chainNetworkName === 'rpc') {
+      const ethAssetHealthCheck = new EthereumRpcAssetHealthCheckParam(
+        ETH,
+        ETH,
+        ethereumContracts.lockAddress,
+        Configs.ethWarnThreshold,
+        Configs.ethCriticalThreshold,
+        GuardsEthereumConfigs.rpc.url,
+        8
+      );
+      healthCheck.register(ethAssetHealthCheck);
+
+      const scannerName = 'ethereum-evm-rpc';
+      const ethereumScannerSyncCheck = new EthereumRPCScannerHealthCheck(
+        generateLastBlockFetcher(scannerName),
+        scannerName,
+        Configs.ethereumScannerWarnDiff,
+        Configs.ethereumScannerCriticalDiff,
+        GuardsEthereumConfigs.rpc.url,
+        GuardsEthereumConfigs.rpc.authToken,
+        GuardsEthereumConfigs.rpc.timeout
+      );
+      healthCheck.register(ethereumScannerSyncCheck);
     }
   }