+++ tests/api/testData.ts
@@ -6,10 +6,18 @@ import { ADA, CARDANO_CHAIN } from '@rosen-chains/cardano';
 import { rosenConfig } from '../../src/configs/RosenConfig';
 import GuardsBitcoinConfigs from '../../src/configs/GuardsBitcoinConfigs';
 import { BITCOIN_CHAIN } from '@rosen-chains/bitcoin';
+import Configs from '../../src/configs/Configs';
 
 export const guardInfo = {
-  version: expect.any(String),
-  health: HealthStatusLevel.HEALTHY,
+  versions: {
+    app: expect.any(String),
+    contract: rosenConfig.contractVersion,
+    tokensMap: Configs.tokensVersion,
+  },
+  health: {
+    status: HealthStatusLevel.HEALTHY,
+    trialErrors: [],
+  },
   balances: {
     hot: [
       {