+++ tests/verification/RequestVerifier.spec.ts
@@ -484,56 +484,6 @@ describe('RequestVerifier', () => {
 
     /**
      * @target RequestVerifier.verifyColdStorageTransactionRequest should return false
-     * when there is already an active cold storage tx for chain in db
-     * @dependencies
-     * - TransactionVerifier
-     * - database
-     * @scenario
-     * - mock transaction
-     * - insert another cold storage transaction into db (same chain)
-     * - mock TransactionVerifier.verifyColdStorageTransaction
-     * - run test
-     * - verify returned value
-     * @expected
-     * - returned value should be false
-     */
-    it('should return false when there is already an active cold storage tx for chain in db', async () => {
-      // mock event and transaction
-      const chain = 'chain';
-      const paymentTx = mockPaymentTransaction(
-        TransactionType.coldStorage,
-        chain,
-        ''
-      );
-
-      // insert another cold storage transaction into db (different chain)
-      const anotherTx = mockPaymentTransaction(
-        TransactionType.coldStorage,
-        chain,
-        ''
-      );
-      await DatabaseActionMock.insertTxRecord(
-        anotherTx,
-        TransactionStatus.approved
-      );
-
-      // mock TransactionVerifier.verifyColdStorageTransaction
-      vi.spyOn(
-        TransactionVerifier,
-        'verifyColdStorageTransaction'
-      ).mockResolvedValue(true);
-
-      // run test
-      const result = await RequestVerifier.verifyColdStorageTransactionRequest(
-        paymentTx
-      );
-
-      // verify returned value
-      expect(result).toEqual(false);
-    });
-
-    /**
-     * @target RequestVerifier.verifyColdStorageTransactionRequest should return false
      * when transaction doesn't satisfy cold storage tx conditions
      * @dependencies
      * - TransactionVerifier